refactor functions in bm25.py and utils.py for modularity
